Business and BusinessSpec: correct available_jobs functionality and test it
I fixed the functionality of the previous version of available_jobs so it now returns an array of available jobs. I tested all of the possible uses that I could think of. I am still not sure about the process of making the result but it seems to get the point across, but I hate that it requires the array class.