Business/BusinessSpec: add self to business methods and tests for delete
Added self to all methods in business so it can access it's jobs (as a business). I also added a .first to the end of the delete return to remove it from the array to make use easier. I also also added tests to flesh out the new uses. (the tests stayed green during the self changes)